# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BTC Portfolio Dashboard API
  version: 1.0.0
  description: |
    REST API for a BTC-only portfolio tracker.

    Authentication uses JWT bearer tokens. Register or log in, then send
    `Authorization: Bearer <token>` for protected endpoints.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://btc-portfolio-backend.onrender.com
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /portfolio/summary:
    get:
      tags: [Portfolio]
      summary: Portfolio summary
      operationId: getPortfolioSummary
      parameters:
        - $ref: "#/components/parameters/CurrencyParam"
      responses:
        "200":
          description: Current portfolio summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortfolioSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /portfolio/history:
    get:
      tags: [Portfolio]
      summary: Portfolio history
      operationId: getPortfolioHistory
      parameters:
        - $ref: "#/components/parameters/CurrencyParam"
        - $ref: "#/components/parameters/RangeParam"
      responses:
        "200":
          description: Portfolio value history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortfolioHistoryPoint"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /portfolio/activity/monthly:
    get:
      tags: [Portfolio]
      summary: Monthly portfolio activity
      operationId: getMonthlyActivity
      responses:
        "200":
          description: Monthly buy/sell activity ordered oldest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyActivity"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    CurrencyParam:
      name: currency
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/Currency"
        default: CHF
    RangeParam:
      name: range
      in: query
      required: false
      schema:
        type: string
        enum: [7d, 30d, 90d, 365d]
        default: 30d
...
  schemas:
...
    Currency:
      type: string
      enum: [CHF, USD, EUR]
...
    PortfolioSummary:
      type: object
      required:
        - currency
        - btc_balance
        - average_cost
        - total_invested
        - net_invested
        - total_fees
        - current_price
        - current_value
        - unrealized_pnl
        - realized_pnl
        - pnl_percent
        - largest_transaction_amount
        - transaction_count
        - first_transaction_date
        - updated_at
      properties:
        currency:
          $ref: "#/components/schemas/Currency"
        btc_balance:
          type: number
        average_cost:
          type: number
        total_invested:
          type: number
        net_invested:
          type: number
        total_fees:
          type: number
        current_price:
          type: number
        current_value:
          type: number
        unrealized_pnl:
          type: number
        realized_pnl:
          type: number
        pnl_percent:
          type: number
        largest_transaction_amount:
          type: number
        transaction_count:
          type: integer
        first_transaction_date:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
    PortfolioHistoryPoint:
      type: object
      required: [timestamp, btc_balance, price, value]
      properties:
        timestamp:
          type: string
          format: date-time
        btc_balance:
          type: number
        price:
          type: number
        value:
          type: number
...
    MonthlyActivity:
      type: object
      required: [month, buy_amount, sell_amount, net_amount, btc_amount, transaction_count]
      properties:
        month:
          type: string
          example: "2026-06"
        buy_amount:
          type: number
        sell_amount:
          type: number
        net_amount:
          type: number
        btc_amount:
          type: number
        transaction_count:
          type: integer
    Error:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ServiceUnavailable:
      description: Market service unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"